Changed movement from UP/LEFT/RIGHT/DOWN-keys to W/S/A/D-movement.
Give this a try, it's perhaps a bit odd to play in an RPG but I think it will fit well.
You can access most of your keys around WSAD (1-0), Questbook, spellbook, character info page etc...

In the end we will have keybindings, however i feel this should be default.